Add support for etek switches in config.json
This is kinda weird and bad since I ended up finding out that codesend did not work with these switches for some reason. I think I could do some more work with timing to get it to work, but not sure. Found a script for the switches, modified it some, and hacked it into this.